refacctored functions into seperate files
